Refactored WebService for improved readability.
