+++ Sources/FlowMagic-SDK/Services/WebService.swift
@@ -16,48 +16,6 @@ protocol WebService {
     func loadUrlData(resource: String) async throws -> ScreenFlowModel
 }
 
-protocol NetworkSession {
-    func loadData(from url: URL) async throws -> (Data, HTTPURLResponse)
-}
-
-extension URLSession: NetworkSession {
-    func loadData(from url: URL) async throws -> (Data, HTTPURLResponse) {
-        let (data, response) = try await self.data(from: url)
-        if let httpResponse = response as? HTTPURLResponse {
-            return (data, httpResponse)
-        } else {
-            throw NetworkError.invalidServerResponse
-        }
-
-    }
-}
-
-extension URLResponse {
-    var isSuccessful: Bool {
-        guard let httpResponse = self as? HTTPURLResponse else {
-            return false
-        }
-
-        return httpResponse.statusCode == 200
-    }
-}
-
-class MockNetworkSession: NetworkSession {
-    var mockData: Data?
-    var mockResponse: HTTPURLResponse?
-    var mockError: Error?
-
-    func loadData(from url: URL) async throws -> (Data, HTTPURLResponse) {
-        if let error = mockError {
-            throw error
-        }
-        guard let data = mockData, let response = mockResponse else {
-            throw NetworkError.invalidServerResponse
-        }
-        return (data, response)
-    }
-}
-
 class WebServiceImpl: WebService {
     func loadUrlData(resource: String) async throws -> ScreenFlowModel {
